backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.6
pinecone-client==3.0.1
sentence-transformers==2.2.2
python-multipart==0.0.6
boto3==1.33.1
google-cloud-storage==2.13.0
reportlab==4.0.4
markdown==3.5.1
requests==2.31.0
aiofiles==23.2.1
